requirements.txt
arrow==0.15.5
boto3==1.11.6
botocore==1.14.6
certifi==2019.11.28
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.1.1
Flask-S3==0.3.3
idna==2.8
imgurpython==1.1.7
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
Pillow==7.0.0
python-dateutil==2.8.1
requests==2.22.0
s3transfer==0.3.1
six==1.14.0
urllib3==1.25.7
Werkzeug==0.16.0
